Split Smart Search file highlighting
Move the post-navigation file highlight intent out of indexer.ts into a dedicated feature-local helper, leaving indexer.ts focused on collecting OPAL search nodes.
